refactor(common): extract changelog markdown renderer
* Extract changelog markdown parsing and rendering into shared ChangelogMarkdown component
* Use previous_version_url to resolve version diffs instead of relying on list order
* Simplify dialog title by deriving previous version label from URL
